chore: remove alpha controller and update core
