+++ pkg/apis/crds/karpenter.sh_nodepools.yaml
@@ -354,6 +354,8 @@ spec:
                         - nodeClassRef
                         - requirements
                       type: object
+                  required:
+                    - spec
                   type: object
                 weight:
                   description: Weight is the priority given to the provisioner during scheduling. A higher numerical weight indicates that this provisioner will be ordered ahead of other provisioners with lower weights. A provisioner with no weight will be treated as if it is a provisioner with a weight of 0.
@@ -361,6 +363,8 @@ spec:
                   maximum: 100
                   minimum: 1
                   type: integer
+              required:
+                - template
               type: object
             status:
               description: NodePoolStatus defines the observed state of NodePool
